Remove TTY tests - dead code (#3229)
