--- cli/tests/is_tty.ts
@@ -1,1 +1,0 @@
-console.log(Deno.isTTY().stdin);
